Update README.md
# TAXUNCHAINED.AI-project

Let's create a system that, given the transaction reports of different exchanges/platforms/crypto wallets, provides the poor Italian with the data to independently compile the tax report. The AI could be free, but also be requested a payment of 50 euros as it is currently necessary to pay accountants hundreds of euros. Milions Italians will buy it for sure.